Add files via upload
The above file, stations-2022-03.10.rsd is the last version that was offered for free download by the RadioSure website before it went offline permanently later that day on 10 March 2022. 

The file you uploaded with a later date, I understand from you, is based on an earlier version of the .rsd file from RadioSure, but it includes updates you did to it as your local copy.

Now that both are here, we need to combine them into a new file that has the changes from each but together in one place.