Add files via upload
File created by converting an .sql database into the .rsd format required by RadioSure application. 
It has very few of the same stations as the final update sent before radiosure.com was shutdown.
It is not as uniform as you might of been used to because the station list is automated and unregulated.